Split Aligner
