+++ SchemeCreator.csproj
@@ -137,6 +137,9 @@
     <Compile Include="Data\Models\WireConnection.cs" />
     <Compile Include="Data\Models\Enums\ConnectionTypeEnum.cs" />
     <Compile Include="Data\Services\Alignment\Aliner.cs" />
+    <Compile Include="Data\Services\Alignment\ExternalOutputsAligner.cs" />
+    <Compile Include="Data\Services\Alignment\ExternalInputsAligner.cs" />
+    <Compile Include="Data\Services\Alignment\GatesAligner.cs" />
     <Compile Include="Data\Services\Colorer.cs" />
     <Compile Include="Data\Services\GateWorkPatterns.cs" />
     <Compile Include="Data\Services\MessageData.cs" />
